More work in the summary
Moved the character image to it's own file
made its size an option in initializer
added image for "no notable raids left"
Started work on previews for widgets for "no raids left"
added character details and image to summary
added toolbar titles to loading and farming summary view
