+++ Shared/Summary/SummaryMain.swift
@@ -13,12 +13,35 @@ struct SummaryMain: View {
         if gameData.loadingAllowed {
             ScrollView {
                 ForEach(gameData.characterRaidEncounters, id: \.character.id) { GDCharacterEncounters in
-                    SingleCharacterSummary(characterEncounters: GDCharacterEncounters)
+                    SingleCharacterSummary(
+                        character: getCharacterBasedOn(encounters: GDCharacterEncounters),
+                        characterEncounters: GDCharacterEncounters
+                    )
                 }
             }
+            .toolbar(content: {
+                ToolbarItem(placement: .principal) {
+                    Text("Farming opportunities:")
+                        .font(.title2)
+                }
+            })
         } else {
             DataLoadingInfo()
+                .toolbar(content: {
+                    ToolbarItem(placement: .principal) {
+                        Text("Loading game data…")
+                            .font(.title3)
+                    }
+                })
         }
+        
+    }
+    
+    func getCharacterBasedOn(encounters: CharacterRaidEncounters) -> CharacterInProfile {
+        let character = gameData.characters.first { (GDCharacter) -> Bool in
+            GDCharacter.name == encounters.character.name && GDCharacter.realm.name == encounters.character.realm.name
+        }
+        return character!
     }
 }
